third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-3/shadow-dom-rules-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
This test checks that style sheets hosted inside shadow roots could be inspected.


Running: testInit

Running: testDumpStyles
[expanded] 
element.style { ()

[expanded] 
.red { (<style>…</style>)
    color: red;

[expanded] 
div { (user agent stylesheet)
    display: block;
